requirements.txt
streamlit>=1.22.0
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.0.0
openpyxl>=3.0.0
pyserial>=3.5
scikit-learn
joblib
ultralytics
opencv-python
requests
flask
flask-login
